Support for reading 24-bit wav files

+++ Source/SFZAudioBuffer.h
@@ -5,6 +5,8 @@
 //  Created by David Wallin on 4/22/13.
 //  Copyright (c) 2013 David Wallin. All rights reserved.
 //
+//  Audio buffer class for loading samples and generating output. Currently only supports
+//  1 or 2 channels.
 
 #ifndef __OpenSFZ__SFZAudioBuffer__
 #define __OpenSFZ__SFZAudioBuffer__
@@ -13,18 +15,18 @@
 class SFZAudioBuffer
 {
 public:
-    SFZAudioBuffer(const int numChannels_, const int numSamples_);
+    SFZAudioBuffer(const int numChannels_, const unsigned long numSamples_);
     
-    SFZAudioBuffer(const int numSamples_, float *channel1, float *channel2);
+    SFZAudioBuffer(const unsigned long numSamples_, float *channel1, float *channel2);
     SFZAudioBuffer(const SFZAudioBuffer &other);
     ~SFZAudioBuffer();
     
     float *channels[2];
     
-    int getNumSamples() { return numSamples; };
+    unsigned long getNumSamples() { return numSamples; };
     int getNumChannels() { return numChannels; };
 private:
-    int numSamples;
+    unsigned long numSamples;
     int numChannels;
     
     bool owned;
